Move tests into test_suite crate
