+++ kinded/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "kinded"
-version = "0.0.1"
+version = "0.0.2"
 edition = "2021"
 authors = ["Serhii Potapov <blake131313@gmail.com>"]
 
@@ -16,4 +16,4 @@ categories = ["data-structures", "rust-patterns"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-kinded_macros = { version = "0.1.0", path = "../kinded_macros" }
+kinded_macros = { version = "0.0.2", path = "../kinded_macros" }
