+++ kinded_macros/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "kinded_macros"
-version = "0.1.0"
+version = "0.0.2"
 edition = "2021"
 authors = ["Serhii Potapov <blake131313@gmail.com>"]
